+++ crates/cast/src/cmd/wallet/vanity.rs
